refactor: remove Validator class and move y/n input validation to inputView class
