+++ review_v2_updates.md
@@ -40,3 +40,4 @@
 - [x] **8. 입력에 trim() 설정하도록 변경**
 - [ ] **9. 모든 파일 포멧팅 최적화**
   - `cmd + opt + L`
+- [ ] **10. validator 없애고, inputView에서 객체가 없는 입력값 확인하도록 변경**
