+++ src/main/java/Application.java
@@ -1,7 +1,6 @@
 import java.util.List;
 import domain.BlackjackGame;
 import java.util.function.Supplier;
-import util.Validator;
 import view.InputView;
 import view.Message;
 import view.OutputView;
@@ -10,12 +9,10 @@ public class Application {
     private final OutputView outputView;
     private final InputView inputView;
     private final BlackjackGame blackjackGame;
-    private final Validator validator;
 
-    public Application(InputView inputView, OutputView outputView, Validator validator, BlackjackGame blackjackGame) {
+    public Application(InputView inputView, OutputView outputView, BlackjackGame blackjackGame) {
         this.inputView = inputView;
         this.outputView = outputView;
-        this.validator = validator;
         this.blackjackGame = blackjackGame;
     }
 
@@ -23,7 +20,6 @@ public class Application {
         new Application(
             new InputView(),
             new OutputView(),
-            new Validator(),
             new BlackjackGame()
         ).run();
     }
@@ -63,7 +59,6 @@ public class Application {
         retryUntilSuccess(() -> {
             outputView.printAskExtraCard(name);
             String answer = inputView.readDealDecision().trim();
-            validator.validateAnswer(answer);
             determinePlayerContinue(answer, index);
             return null;
         });
