--- src/main/java/util/Validator.java
@@ -1,22 +1,0 @@
-package util;
-
-public class Validator {
-    public void validateAnswer(String answer) {
-        validateGetMoreEmptyInput(answer);
-        validateYesOrNo(answer);
-    }
-    
-    static void validateGetMoreEmptyInput(String answer) {
-        if (answer.trim().isEmpty()) {
-            throw new IllegalArgumentException("[ERROR] 카드 수령 여부(y/n)는 비어 있을 수 없습니다.");
-        }
-    }
-
-    static void validateYesOrNo(String answer) {
-        if (!(answer.equalsIgnoreCase("y") || answer.equalsIgnoreCase("n"))) {
-            throw new IllegalArgumentException("[ERROR] 'y' 또는 'n'만 입력해 주세요.");
-        }
-    }
-
-
-}
