+++ src/test/java/ApplicationTest.java
@@ -10,7 +10,6 @@ import java.util.List;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import util.Validator;
 import view.InputView;
 import view.OutputView;
 import vo.Rank;
@@ -48,7 +47,7 @@ public class ApplicationTest {
 
         Application app = new Application(
                 new InputView(), new OutputView(),
-                new Validator(), new TestBlackjackGame(new Deck(FIXED_CARDS))
+                new TestBlackjackGame(new Deck(FIXED_CARDS))
         );
 
         // when
@@ -69,7 +68,7 @@ public class ApplicationTest {
 
         Application app = new Application(
                 new InputView(), new OutputView(),
-                new Validator(), new TestBlackjackGame(new Deck(FIXED_CARDS))
+                new TestBlackjackGame(new Deck(FIXED_CARDS))
         );
 
         // when
@@ -88,7 +87,7 @@ public class ApplicationTest {
 
         Application app = new Application(
                 new InputView(), new OutputView(),
-                new Validator(), new TestBlackjackGame(new Deck(FIXED_CARDS))
+                new TestBlackjackGame(new Deck(FIXED_CARDS))
         );
 
         // when
